requirements.txt
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
pytest==8.3.4
streamlit==1.43.2
